requirements.txt
zxing-cpp
pillow
qrcode